Extract TestWarPackagingContext to fix checkstyle MethodLength violation
